fix for windows if the downloaded binary .exe
folders are a cibling to the php file using it, the "windows-example-project" shows an example of how this could look along with the downloaded binaries from the site.  ctrl-f (find)  "*_W" to find changes to code. This is a quick fix to the library to work with my project on Windows 10 / Windows Server